changed: move non-template functions in BoundaryPeriodicity to a separate compilation unit
