+++ CMakeLists_files.cmake
@@ -35,6 +35,7 @@ list (APPEND MAIN_SOURCE_FILES
 	opm/porsol/blackoil/fluid/MiscibilityLiveOil.cpp
 	opm/porsol/blackoil/fluid/MiscibilityProps.cpp
 	opm/porsol/common/blas_lapack.cpp
+	opm/porsol/common/BoundaryPeriodicity.cpp
 	opm/porsol/common/LinearSolverISTL.cpp
 	)
